Fix for new btrfs tools/kernel/etc. in 14.04
With the current kernel and btrfs tools in Ubuntu 14.04, newly-deleted
snapshots or subvolumes apparently appear in the output of 'btrfs
subvolume list' with a path of 'DELETED'. These were identified as
subvolumes that we should be snapshotted, which causes a boot failure
and a '(initramfs)' BusyBox prompt since 'DELETED' isn't the actual path
to a subvolume on disk.

Ignore these when considering subvolumes that should be snapshotted.